ajout de la dependance cloud firestore et commencer a paramétré l'enregistrement des dépense en base de données
