+++ pubspec.yaml
@@ -33,6 +33,7 @@ dependencies:
 
   firebase_core: ^2.24.2
   firebase_auth: ^4.17.5
+  cloud_firestore: ^4.15.10
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
